requirements.txt
aws-cdk-lib==2.143.0
constructs>=10.0.0,<11.0.0
requests==2.32.2
aws-cdk.aws-lambda-python-alpha==2.143.0a0
